[Feat] Divide extension feature to other package.
