+++ ahttp_client/extension/pydantic.py
@@ -1,114 +1,38 @@
+
+"""MIT License
+
+Copyright (c) 2023 gunyu1019
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+"""
 from __future__ import annotations
 
 import inspect
-from collections.abc import Sequence
-from types import GenericAlias
-from typing import overload, TYPE_CHECKING, TypeVar
-
 import aiohttp
 
-T = TypeVar("T")
+from collections.abc import Sequence
+from types import GenericAlias
+from typing import overload, TypeVar, TYPE_CHECKING
 
 if TYPE_CHECKING:
-    import asyncio
+    from typing import Any, Optional
 
-    from abc import ABC, abstractmethod
-    from typing import Any, Callable, Generic, Optional
-    from ._types import RequestAfterHookFunction, RequestBeforeHookFunction
-    from .query import Query
-    from .request import RequestCore, request
-    from .session import Session
-
-    CallableT = TypeVar("CallableT")
-    CallableR = TypeVar("CallableR")
-
-    class BoundedMethod(ABC, Generic[T, CallableT, CallableR]):
-        @abstractmethod
-        @property
-        def __self__(self) -> T: ...
-
-        @abstractmethod
-        @property
-        def __func__(self) -> Callable[[T, CallableT], CallableR]: ...
-
-    MultipleHookT = BoundedMethod[T, CallableT, CallableR] | Callable[[CallableT], CallableR]
-
-
-def multiple_hook(
-    hook: MultipleHookT[
-        RequestCore,
-        [RequestAfterHookFunction | RequestBeforeHookFunction],
-        RequestAfterHookFunction | RequestBeforeHookFunction,
-    ],
-    index: Optional[int] = -1,
-):
-    """Use this method, if more than one pre-invoke hooks or post-invoke hooks need.
-
-    Parameters
-    ----------
-    hook: Callable[
-            [RequestAfterHookFunction | RequestBeforeHookFunction],
-            RequestAfterHookFunction | RequestBeforeHookFunction
-        ]
-        Contains the decorator function used for hooking.
-        This can be :meth:`RequestObj.before_hook` or :meth:`RequestObj.after_hook`.
-    index: Optional[int]
-        Order of invocation in invoke-hook
-
-    Warnings
-    --------
-    This feature is experimental. It might not work as expected.
-
-    Examples
-    --------
-    >>> class MetroAPI(Session):
-    ...    def __init__(self, loop: asyncio.AbstractEventLoop):
-    ...        super().__init__("https://api.yhs.kr", loop=loop)
-    ...
-    ...    @request("GET", "/metro/station")
-    ...    async def station_search_with_query(
-    ...            self,
-    ...            response: aiohttp.ClientResponse,
-    ...            name: Query | str
-    ...    ) -> dict[str, Any]:
-    ...        return await response.json()
-    ...
-    ...    @multiple_hook(station_search_with_query.before_hook)
-    ...    async def before_hook_1(self, obj, path):
-    ...        # Set-up before request
-    ...        return obj, path
-    ...
-    ...    @multiple_hook(station_search_with_query.before_hook)
-    ...    async def before_hook_2(self, obj, path):
-    ...        # Set-up before request
-    ...        return obj, path
-    """
-    hook_name = hook.__name__
-    instance = hook.__self__
-
-    is_exist_multiple_hook = True
-    if not hasattr(instance, "__multiple_%s__" % hook_name):
-        is_exist_multiple_hook = False
-        setattr(instance, "__multiple_%s__" % hook_name, [])
-
-    def decorator(func):
-        getattr(instance, "__multiple_%s__" % hook_name).append((func, index))
-        return func
-
-    async def wrapper(wrapped_instance, *args):
-        _args = tuple(args)
-        multiple_hook_func = getattr(instance, "__multiple_%s__" % hook_name)
-        sorted_multiple_hook_func = sorted(multiple_hook_func, key=lambda x: x[1])
-        for func, _ in sorted_multiple_hook_func:
-            _args = await func(wrapped_instance, *_args)
-        return _args
-
-    if not is_exist_multiple_hook:
-        hook.__func__(instance, wrapper)
-    return decorator
-
-
-# Pydantic Model Validate
 try:
     import pydantic
 except (ModuleNotFoundError, ImportError):
@@ -122,7 +46,7 @@ else:
 @overload
 def _parsing_json_to_model(
     data: list[Any],
-    model: BaseModelT,
+    model: type[BaseModelT],
     /,
     *,
     strict: Optional[bool] = None,
@@ -133,7 +57,7 @@ def _parsing_json_to_model(
 
 def _parsing_json_to_model(
     data: dict[Any, ...],
-    model: BaseModelT,
+    model: type[BaseModelT],
     *,
     strict: Optional[bool] = None,
     from_attributes: Optional[bool] = None,
